Move slots to basic
