+++ packages/integrations/mdx/test/fixtures/basic/src/components/slots/Slotted.astro
