+++ packages/integrations/mdx/test/fixtures/basic/src/components/slots/Test.mdx
