+++ packages/integrations/mdx/test/fixtures/basic/src/pages/slots/glob.astro
@@ -1,5 +1,5 @@
 ---
-const components = await Astro.glob('../components/*.mdx');
+const components = await Astro.glob('../../components/slots/*.mdx');
 ---
 
 <div data-default-export>
