+++ packages/integrations/mdx/test/fixtures/basic/src/pages/slots/index.astro
@@ -1,0 +1,5 @@
+---
+import Test from '../../components/slots/Test.mdx';
+---
+
+<Test />
