--- packages/integrations/mdx/test/fixtures/mdx-slots/src/pages/index.astro
@@ -1,5 +1,0 @@
----
-import Test from '../components/Test.mdx';
----
-
-<Test />
